64 bit fix
git-svn-id: svn://svn.savannah.gnu.org/mtools/trunk@202 8fbfe7c3-7cdf-472d-ae67-d119146b2a64

+++ init.c
@@ -308,7 +308,7 @@ Stream_t *fs_init(char drive, int mode, int *isRop)
 	if(!mtools_skip_check && (tot_sectors % dev.sectors)){
 		fprintf(stderr,
 			"Total number of sectors (%d) not a multiple of"
-			" sectors per track (%d)!\n", tot_sectors,
+			" sectors per track (%d)!\n", (int) tot_sectors,
 			dev.sectors);
 		fprintf(stderr,
 			"Add mtools_skip_check=1 to your .mtoolsrc file "
